third_party/blink/web_tests/virtual/plz-service-worker/http/tests/serviceworker/register-error-messages-expected.txt

This is a testharness.js-based test.
FAIL Registering same scope as the script directory without the last slash assert_equals: expected "Failed to register a ServiceWorker for scope ('http://127.0.0.1:8000/serviceworker/resources') with script ('http://127.0.0.1:8000/serviceworker/resources/registration-worker.js'): The path of the provided scope ('/serviceworker/resources') is not under the max scope allowed ('/serviceworker/resources/'). Adjust the scope, move the Service Worker script, or use the Service-Worker-Allowed HTTP header to allow the scope." but got "Failed to register a ServiceWorker for scope ('http://127.0.0.1:8000/serviceworker/resources') with script ('http://127.0.0.1:8000/serviceworker/resources/registration-worker.js'): Operation failed by security issue"
FAIL Registration scope outside the script directory assert_equals: expected "Failed to register a ServiceWorker for scope ('http://127.0.0.1:8000/serviceworker/different-directory/') with script ('http://127.0.0.1:8000/serviceworker/resources/registration-worker.js'): The path of the provided scope ('/serviceworker/different-directory/') is not under the max scope allowed ('/serviceworker/resources/'). Adjust the scope, move the Service Worker script, or use the Service-Worker-Allowed HTTP header to allow the scope." but got "Failed to register a ServiceWorker for scope ('http://127.0.0.1:8000/serviceworker/different-directory/') with script ('http://127.0.0.1:8000/serviceworker/resources/registration-worker.js'): Operation failed by security issue"
PASS Registration scope outside domain
PASS Registering script outside domain
FAIL Registering non-existent script assert_equals: expected "Failed to register a ServiceWorker for scope ('http://127.0.0.1:8000/serviceworker/resources/scope/no-such-worker') with script ('http://127.0.0.1:8000/serviceworker/resources/no-such-worker.js'): A bad HTTP response code (404) was received when fetching the script." but got "Failed to register a ServiceWorker for scope ('http://127.0.0.1:8000/serviceworker/resources/scope/no-such-worker') with script ('http://127.0.0.1:8000/serviceworker/resources/no-such-worker.js'): Operation failed by network issue"
PASS Registering invalid chunked encoding script
FAIL Registering script with no MIME type assert_equals: expected "Failed to register a ServiceWorker for scope ('http://127.0.0.1:8000/serviceworker/resources/scope/no-mime-type-worker/') with script ('http://127.0.0.1:8000/serviceworker/resources/mime-type-worker.php'): The script does not have a MIME type." but got "Failed to register a ServiceWorker for scope ('http://127.0.0.1:8000/serviceworker/resources/scope/no-mime-type-worker/') with script ('http://127.0.0.1:8000/serviceworker/resources/mime-type-worker.php'): Operation failed by security issue"
FAIL Registering script with bad MIME type assert_equals: expected "Failed to register a ServiceWorker for scope ('http://127.0.0.1:8000/serviceworker/resources/scope/bad-mime-type-worker/') with script ('http://127.0.0.1:8000/serviceworker/resources/mime-type-worker.php?mime=text/plain'): The script has an unsupported MIME type ('text/plain')." but got "Failed to register a ServiceWorker for scope ('http://127.0.0.1:8000/serviceworker/resources/scope/bad-mime-type-worker/') with script ('http://127.0.0.1:8000/serviceworker/resources/mime-type-worker.php?mime=text/plain'): Operation failed by security issue"
FAIL Registering redirected script assert_equals: expected "Failed to register a ServiceWorker for scope ('http://127.0.0.1:8000/serviceworker/resources/scope/redirect/') with script ('http://127.0.0.1:8000/serviceworker/resources/redirect.php?Redirect=..%2F%2Fresources%2Fregistration-worker.js'): The script resource is behind a redirect, which is disallowed." but got "Failed to register a ServiceWorker for scope ('http://127.0.0.1:8000/serviceworker/resources/scope/redirect/') with script ('http://127.0.0.1:8000/serviceworker/resources/redirect.php?Redirect=..%2F%2Fresources%2Fregistration-worker.js'): Operation failed by security issue"
PASS Registering script including parse error
PASS Registering script including undefined error
PASS Registering script including uncaught exception
PASS Registering script importing malformed script
PASS Registering script importing non-existent script
PASS Script URL including URL-encoded slash
Harness: the test ran to completion.
